fix:refactoring to atomic design
