+++ src/components/organisms/CommunityShowcase.tsx
@@ -1,4 +1,4 @@
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/atoms/Card'
 
 const communities = [
   {
