+++ src/components/organisms/Experience.tsx
@@ -1,7 +1,7 @@
 import { Briefcase } from 'lucide-react'
 
-import { Badge } from '@/components/ui/badge'
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
+import { Badge } from '@/components/atoms/Badge'
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/atoms/Card'
 
 const experiences = [
   {
