+++ src/components/organisms/index.ts
@@ -1,0 +1,9 @@
+export { Certifications } from './Certifications'
+export { CommunityShowcase } from './CommunityShowcase'
+export { Contact } from './Contact'
+export { Experience } from './Experience'
+export { Hero } from './Hero'
+export { Navigation } from './Navigation'
+export { Projects } from './Projects'
+export { Skills } from './Skills'
+export { SocialLinks } from './SocialLinks'
