+++ src/index.css
@@ -72,7 +72,7 @@
     font-family: 'Aoboshi One', 'Noto Sans JP', 'Hiragino Kaku Gothic ProN', 'Yu Gothic', 'Meiryo', sans-serif;
     font-feature-settings: 'palt' 1;
     font-weight: 500;
-    background-image: url('./components/img/wallpaper.png');
+    background-image: url('./assets/images/wallpaper.png');
     background-size: cover;
     background-position: center;
     background-repeat: no-repeat;
