+++ src/main.tsx
@@ -1,10 +1,11 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+
 import './index.css'
-import App from './App.tsx'
+import HomePage from '@/pages/home/HomePage'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+    <HomePage />
   </StrictMode>,
 )
